refactor(core): update imports and namespaces
