+++ src/test/java/dev/persiantools/BankTest.java
@@ -1,6 +1,6 @@
-package com.github.persian.tools;
+package dev.persiantools;
 
-import com.github.persian.usecases.exceptions.BankNotFoundByProvidedDigits;
+import dev.persiantools.usecases.exceptions.BankNotFoundByProvidedDigits;
 import org.junit.Test;
 
 import java.util.HashMap;
